fix(selfie): fix /selfie on Windows — camera overlay, native field visibility, correct JPEG output
- Hide native text-field HWNDs (room search, compose bar) while the camera
  overlay is open, using the new camera_overlay_open() accessor on MainAppWidget
- Add MFStartup/MFShutdown in main.cpp as the single process-lifetime Media
  Foundation owner; remove ad-hoc ensure_mf_started() calls from audio_win32,
  video_win32, and video_decode_mf
- Switch video_capture_win32 from MF_SOURCE_READER_ENABLE_VIDEO_PROCESSING to
  MF_SOURCE_READER_ENABLE_ADVANCED_VIDEO_PROCESSING: the Color Converter DSP
  (basic flag) does not support ARGB32 output, so SetCurrentMediaType(ARGB32)
  silently fell back to YUY2; the Video Processor MFT (advanced flag) does
  support ARGB32, delivering correct BGRA frames with alpha=255
- Revert CameraWidget to set_bgra_callback with a plain buffer copy, replacing
  the manual BT.601 I420→BGRA loop that was added as a workaround
- Replace manual BGRA→BGR alpha-strip loop in the WIC JPEG encoder with
  IWICBitmap + IWICFormatConverter::Initialize + WriteSource

+++ ui/shared/views/CameraWidget.cpp
@@ -35,7 +35,7 @@ void CameraWidget::open()
         [this](const std::uint8_t* bgra, std::uint32_t w, std::uint32_t h)
         {
             std::lock_guard<std::mutex> lk(frame_mu_);
-            last_bgra_.assign(bgra, bgra + w * h * 4);
+            last_bgra_.assign(bgra, bgra + static_cast<std::size_t>(w) * h * 4);
             frame_w_ = w;
             frame_h_ = h;
         });
